add shared types package between crds
