+++ helm/ngrok-operator/templates/agent/rbac.yaml
@@ -54,6 +54,25 @@ rules:
   - patch
   - update
 - apiGroups:
+  - ngrok.k8s.ngrok.com
+  resources:
+  - ngroktrafficpolicies
+  verbs:
+  - get
+  - list
+  - watch
+- apiGroups:
+  - ingress.k8s.ngrok.com
+  resources:
+  - domains
+  verbs:
+  - get
+  - list
+  - watch
+  - create
+  - patch
+  - delete
+- apiGroups:
   - ""
   resources:
   - events
